fix(icon-button): include typescript support
