+++ packages/react/src/components/IconButton/index.ts
@@ -1,0 +1,3 @@
+export type { IconButtonProps } from './IconButton';
+
+export { IconButton } from './IconButton';
